==> ml/4_detect.ipynb <==
import subprocess
import time
import datetime
import pandas as pd
from pathlib import Path

def get_wifis():
    subprocess.run(
        ['netsh', 'interface', 'set', 'interface', 'name="Wi-Fi"', 'admin=disabled'],
        capture_output=True)
    subprocess.run(['netsh', 'interface', 'set', 'interface', 'name="Wi-Fi"', 'admin=enabled'],
        capture_output=True)

    time.sleep(3)
    output = subprocess.run(
        ['netsh', 'wlan', 'show', 'network', 'mode=Bssid'],
        capture_output=True, text=True, encoding='ISO-8859-1').stdout

    results = output.split('\n\n')[1:-1]

    timestamp = datetime.datetime.now()
    wifis = []
    for result in results:
        lines = result.split('\n')

        for i in range(len(lines)):
            if lines[i].split()[0] == 'BSSID':
                bssid = lines[i].split()[-1]
                rssi = int(lines[i+1].split()[-1][:-1])
                if rssi > 50:
                    wifis.append({'bssid':bssid, 'rssi':rssi, 'timestamp':timestamp})

    return wifis

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import joblib
from collections import defaultdict
from sklearn.preprocessing import LabelEncoder

position = input('enter position : ')
model_path = Path('../ml_model') / position

model_rdf = joblib.load(model_path / 'model_rdf.plk')
model_svm = joblib.load(model_path / 'model_svm.plk')

bssids = np.load(model_path / 'features.npy')
bssids

rp_encoder = LabelEncoder()
rp_encoder.classes_ = np.load(model_path / 'classes.npy', allow_pickle=True)
rp_encoder.classes_

user_wifis = get_wifis()

wifi_dict = defaultdict.fromkeys(bssids, 0)
for user_wifi in user_wifis:
    if user_wifi['bssid'] in wifi_dict:
        wifi_dict[user_wifi['bssid']] = user_wifi['rssi']
        
np.asarray([rssi for bssid, rssi in wifi_dict.items()])

input_df = pd.DataFrame.from_dict([dict(wifi_dict)])
input_df


pred_rdf = model_rdf.predict(input_df)
pred_svm = model_svm.predict(input_df)

rdf_rp = rp_encoder.inverse_transform(pred_rdf)[0]
svm_rp = rp_encoder.inverse_transform(pred_svm)[0]

print(f'rdf : {rdf_rp}, svm : {svm_rp}')

rdf_pred = model_rdf.predict_proba(input_df)[0]
svm_pred = model_svm.predict_proba(input_df)[0]

sorted(list(zip(rdf_pred, rp_encoder.classes_)), key=lambda p:p[0], reverse=True)

sorted(list(zip(svm_pred, rp_encoder.classes_)), key=lambda p:p[0], reverse=True)

ave_pred = [(p1+p2)/2 for p1,p2 in list(zip(rdf_pred, svm_pred))]
ave_pred_rp = list(zip(ave_pred, rp_encoder.classes_))
ave_pred_rp_sorted = sorted(ave_pred_rp, key=lambda p: p[0], reverse=True)
ave_pred_rp_sorted

import sklearn
sklearn.__version__

import joblib
joblib.__version__
